pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "truth"
version = "0.1.0"
description = "An open-source library to solve oracle problems using LLMs, Brave Search, and UMA"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["llm", "oracle", "verification", "brave search", "uma"]
authors = [
  { name = "Murat", email = "murat@2lw.ai" },
  { name = "Efe", email = "efe@2lw.ai" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
]
dependencies = [
  "mistralai",
  "requests",
  "python-dotenv",
]

[project.urls]
Homepage = "https://github.com/yourusername/truth"
Repository = "https://github.com/yourusername/truth.git"
"Bug Tracker" = "https://github.com/yourusername/truth/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/truth"]